Fixed another bug in the database creation SQL script.

+++ sql/ZodiacCipher.sql
@@ -9,7 +9,6 @@ CREATE DATABASE "Zodiac"
        LC_COLLATE = 'English, United States'
        LC_CTYPE = 'English, United States'
        CONNECTION LIMIT = -1;
-
 	   
 -- Table: parts_of_speech
 
@@ -19,13 +18,14 @@ CREATE TABLE parts_of_speech
 (
   word character varying NOT NULL,
   part_of_speech character(1) NOT NULL,
-  frequency_weight smallint DEFAULT 1,
-  CONSTRAINT pk_word_pos PRIMARY KEY (word, part_of_speech)
+  frequency_weight integer DEFAULT 1,
+  CONSTRAINT pk_word_pos PRIMARY KEY (word , part_of_speech )
 )
 WITH (
   OIDS=FALSE
 );
-ALTER TABLE parts_of_speech OWNER TO postgres;
+ALTER TABLE parts_of_speech
+  OWNER TO postgres;
 	   
 
 -- Table: cipher
